+++ device_inductance/utils.py
@@ -1,8 +1,14 @@
 from datetime import datetime
-from typing import TypeVar, Iterator
+from typing import TypeVar, Iterator, Optional, Callable
 
 import numpy as np
 from numpy.typing import NDArray
+from scipy.constants import mu_0
+
+from scipy.sparse import csc_matrix
+from scipy.sparse.linalg import factorized
+
+from cfsem import gs_operator_order4, flux_circular_filament
 
 
 T = TypeVar("T")
@@ -172,3 +178,75 @@ def calc_flux_density_from_flux(
     bz = r_inv * dpsidr / (2.0 * np.pi)  # [T]
 
     return (br, bz)
+
+
+def flux_solver(
+    grids: tuple[NDArray, NDArray], meshes: tuple[NDArray, NDArray]
+) -> Callable[[NDArray], NDArray]:
+    """
+    Linear solver for extracting a flux field from a toroidal current density distribution
+    using a 4th-order finite difference approximation of the Grad-Shafranov PDE.
+    For `jtor` toroidal current density shaped like (nr, nz), call like `psi = flux_solver(rhs)`
+    to get `psi` in [Wb] or [V-s], where `rhs = -2.0 * np.pi * mu_0 * rmesh * jtor` with the boundary
+    values set to the circular-filament solved flux.
+    """
+    # Build Grad-Shafranov Delta* linear operator for finite difference
+    # as a sparse matrix
+    _ = _check_regular(grids)
+    nr, nz = meshes[0].shape
+    vals, rows, cols = gs_operator_order4(*grids)
+    operator = csc_matrix((vals, (rows, cols)), shape=(nr * nz, nr * nz))
+    # Store LU factorization of operator matrix to allow fast, reusable
+    # solves using different right-hand-side (different current density)
+    return factorized(operator)
+
+
+def solve_flux_axisymmetric(
+    grids: tuple[NDArray, NDArray],
+    meshes: tuple[NDArray, NDArray],
+    current_density: NDArray,
+    solver: Optional[Callable[[NDArray], NDArray]],
+) -> NDArray:
+    # Build the differential operator, if needed
+    solver = solver or flux_solver(grids, meshes)
+
+    # Unpack and filter down to just useful inputs
+    dr, dz = _check_regular(grids)  # [m] grid spacing
+    area = dr * dz  # [m^2]
+    rmesh, zmesh = meshes  # [m]
+    nonzero_inds = np.where(current_density != 0.0)
+    current_density_nonzero = np.ascontiguousarray(
+        current_density[nonzero_inds]
+    )  # [A/m^2]
+    rmesh_nonzero = np.ascontiguousarray(rmesh[nonzero_inds])  # [m]
+    zmesh_nonzero = np.ascontiguousarray(zmesh[nonzero_inds])  # [m]
+    # Solve `Delta* @ psi = -mu_0 * 2pi * rmesh * jtor`
+    #   Set up right-hand-side of Grad-Shafranov
+    rhs = -(2.0 * np.pi * mu_0) * rmesh * current_density  # [Wb/m^2]
+    #   Set flux boundary condition
+    #   For most relevant grid sizes (up to 500 X 500), doing the O(N^3)
+    #   circular-filament flux calc is faster than the linear solve
+    #   and therefore faster than doing an extra fixed-boundary linear solve
+    #   in order to use Von Hagenow's asymptotically-O(N^2logN) method.
+    ifil = (area * current_density_nonzero).flatten()  # [A] plasma filament current
+    rfil = rmesh_nonzero.flatten()
+    zfil = zmesh_nonzero.flatten()
+    for s in [[0, ...], [-1, ...], [..., 0], [..., -1]]:  # All boundary slices
+        rhs[s[0], s[1]] = flux_circular_filament(
+            ifil, rfil, zfil, rmesh[s[0], s[1]], zmesh[s[0], s[1]]
+        )
+    #   Do the actual linear solve
+    psi = solver(rhs.flatten()).reshape(rmesh.shape)  # [Wb]
+
+    return psi
+
+
+def _check_regular(grids: tuple[NDArray, NDArray]) -> tuple[float, float]:
+    """Check that grids are regular and returns spacing"""
+    rgrid, zgrid = grids
+    drs = np.diff(rgrid)
+    dzs = np.diff(zgrid)
+    assert np.all(drs == drs[0]), "Grids must be regular"
+    assert np.all(dzs == dzs[0]), "Grids must be regular"
+
+    return drs[0], dzs[0]  # [m]