+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Added
 
-* 
+* `utils.flux_solver` and `utils.solve_flux_axisymmetric` functions extracted from device for solving continuous flux fields by finite difference
+* `coils.Coil.[grids, meshes, local_flux_table]` cached properties for solving a smooth self-field over coil winding packs if they can be mapped on to a regular grid
+
+### Changed
+
+* Plasma flux functions defer to new flux solver functions in `utils`
 
 ## 1.7.4 - 2024-10-09